WIP Makefile: add a NO_INSTALL_CP_FALLBACK flag
Change the behavior of "make install", which will try hardlinking by
default, to optionally distinguish the failure case where it falls
back to copying because it's been told so by a NO_*_HARDLINKS option,
and where it falls back because those options weren't provided but
hardlinking failed.

This is part of making the installation procedure optionally
stricter. The reason for why these fallbacks are here is because on
some OSs or mountpoint combinations we may not be able to hardlink or
symlink, or hardlink across directories, that's why we have the
NO_*_HARDLINKS settings.

But someone who knows they're e.g. building on a well-functioning *nix
system might want to know if they're somehow in a state where
hardlinking or symlinking doesn't work, rather than silently falling
back on producing a git installation that's going to be much bigger
since everything is copied instead of linked.

FIXME: The topmost case needs to be amended to work with this as well,
but it's more complex because we have both NO_INSTALL_HARDLINKS and
NO_CROSS_DIRECTORY_HARDLINKS.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>