worktree: be clearer when "add" dwim-ery kicks in
Currently there is no indication in the "git worktree add" output
that a new branch was created.  This would be especially useful
information in the case where the dwim of "git worktree add <path>"
kicks in, as the user didn't explicitly ask for a new branch, but we
create one for them.

Print some additional output showing that a branch was created and
the branch name to help the user.

This will also be useful in a later commit, which introduces a new
kind of dwim-ery of checking out the branch if it exists instead of
refusing to create a new worktree in that case, and where it's nice
to tell the user which kind of dwim-ery kicked in.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>