fetch tests: add scaffolding for testing a clobbered tag
Add scaffolding for testing how tag clobbering works in combination
with fetch. This is being added to the increasingly inaccurately named
test_configured_prune_type() function because it already tests all the
hairy edge cases, and it's easier to just add another parameter to it
and benefit from the existing tests.

See my e249ce0ccd ("fetch tests: add scaffolding for the new
fetch.pruneTags", 2018-02-09) for another such scaffolding
commit. This will be used by a subsequent change to change the
behavior of tag clobbering on fetch.

The ${10} syntax is needed because some POSIX shells will error out if
the tenth argument is not quoted like this. See my
de8bada2bf ("wildmatch test: create & test files on disk in addition
to in-memory", 2018-01-30) for prior art.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>