git-subtree: move from contrib/subtree/
Change the git-subtree command from living in contrib to being a
"real" git command.

I think it's useful enough that it should be shipped by default, and
unlike git-submodule doesn't need any changes to git's object
format. It's just a helper to do things you can do manually already
with plumbing tools (but that would be much more tedious).

It'll also help to avoid the confusion noted in
https://public-inbox.org/git/alpine.LFD.2.21.1802070547040.5530@android-a172fe96dd584b41/
where users know git-subtree exists, but can't find its
documentation (or it isn't installed).

This change removes config/subtree's copy of COPYING because it's
redundant to the one in the root of the tree. The contrib/subtree/todo
file is being removed because we don't usually store TODOs
in-tree (and this'll be available in the history), the rest like the
.gitignore, Makefile etc. are all redundant to things we have in-tree.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>