doc revisions / tutorial: stop inline attributing Jon Loeliger
These were added way back in 2005 in 66158e331b ("Illustration:
"Fundamental Git Index Operations"", 2005-11-03) and
2be8fd085e ("Illustration: "Commit DAG Revision Naming"",
2005-11-03).

We're no longer attributing every bit of ASCII art in the
documentation to the relevant author in the documentation, and it's
distracting to the reader to make an exception here. We'll still be
attributing the author in the log.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>